refactor: 역할에 맞게 클래스 패키지 이동 및 구조 정리
