+++ src/main/java/janggi/model/initializer/BoardInitializer.java
@@ -1,8 +1,9 @@
 package janggi.model.initializer;
 
-import janggi.model.Board;
+import janggi.model.board.Board;
 import janggi.model.gimul.AbstractGimul;
 import janggi.model.position.Position;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -37,4 +38,42 @@ public abstract class BoardInitializer {
     protected abstract Map<Position, AbstractGimul> initByeong();
 
     protected abstract Map<Position, AbstractGimul> initPho();
+
+    public enum BoardType {
+        LEFT(1) {
+            public Board init() {
+                return new LeftSidedTableSetting().init();
+            }
+        },
+        RIGHT(2) {
+            public Board init() {
+                return new RightSidedTableSetting().init();
+            }
+        },
+        INSIDE(3) {
+            public Board init() {
+                return new InsideTableSetting().init();
+            }
+        },
+        OUTSIDE(4) {
+            public Board init() {
+                return new OutsideTableSetting().init();
+            }
+        };
+
+        private final int number;
+
+        BoardType(int number) {
+            this.number = number;
+        }
+
+        public abstract Board init();
+
+        public static BoardType of(int number) {
+            return Arrays.stream(values())
+                    .filter(type -> type.number == number)
+                    .findFirst()
+                    .orElseThrow(() -> new IllegalArgumentException("유효한 유형 번호를 입력하세요."));
+        }
+    }
 }
