+++ src/test/java/janggi/BoardTest.java
@@ -3,9 +3,9 @@ package janggi;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import janggi.model.Board;
 import janggi.model.Score;
 import janggi.model.Team;
+import janggi.model.board.Board;
 import janggi.model.gimul.AbstractGimul;
 import janggi.model.gimul.diagonalMove.Ma;
 import janggi.model.gimul.linearMove.Cha;
