+++ src/test/java/janggi/gimul/ByeongTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import janggi.model.Score;
 import janggi.model.Team;
 import janggi.model.gimul.AbstractGimul;
-import janggi.model.gimul.Byeong;
+import janggi.model.gimul.byeong.Byeong;
 import janggi.model.gimul.linearMove.Cha;
 import janggi.model.position.Column;
 import janggi.model.position.Position;
