+++ src/test/java/janggi/turn/HanTurnTest.java
@@ -2,8 +2,8 @@ package janggi.turn;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import janggi.model.Board;
 import janggi.model.Team;
+import janggi.model.board.Board;
 import janggi.model.gimul.AbstractGimul;
 import janggi.model.gimul.diagonalMove.Ma;
 import janggi.model.gimul.palace.Jang;
